requirements.txt
asn1crypto==0.24.0
cffi==1.12.2
cryptography==2.6.1
Django==2.1.8
django-cors-headers==2.5.2
django-rest-knox==4.0.1
djangorestframework==3.9.2
djangorestframework-simplejwt==4.3.0
Pillow==6.0.0
pycparser==2.19
PyJWT==1.7.1
pytz==2018.9
six==1.12.0
sqlparse==0.3.0
